test(world): Move TestingVoxelWorld and related classes to world package
