+++ src/test/java/com/ignfab/minalac/generator/generation/TestingGenerationTile.java
@@ -3,10 +3,10 @@ package com.ignfab.minalac.generator.generation;
 import com.ignfab.minalac.generator.generation.heightmaps.HeightmapDeclarationStore;
 import com.ignfab.minalac.generator.generation.heightmaps.HeightmapStore;
 import com.ignfab.minalac.generator.generation.heightmaps.TestingHeightmap;
-import com.ignfab.minalac.generator.outputs.testing.TestingVoxelTile;
-import com.ignfab.minalac.generator.outputs.testing.TestingVoxelWorld;
 import com.ignfab.minalac.generator.utils.world2d.WorldBBox2d;
 import com.ignfab.minalac.generator.utils.world3d.WorldBBox3d;
+import com.ignfab.minalac.generator.world.TestingVoxelTile;
+import com.ignfab.minalac.generator.world.TestingVoxelWorld;
 
 /**
  * A fake GenerationTile over a {@link TestingVoxelWorld}.
