+++ src/test/java/com/ignfab/minalac/generator/generation/heightmaps/computed/operators/CappedManhattanHeightmapOperatorTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 
 import com.ignfab.minalac.generator.generation.heightmaps.Heightmap;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class CappedManhattanHeightmapOperatorTest {
     @Test
