+++ src/test/java/com/ignfab/minalac/generator/models/filters/ModelFilterOnMetadataValueTest.java
@@ -2,7 +2,7 @@ package com.ignfab.minalac.generator.models.filters;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class ModelFilterOnMetadataValueTest {
     @Test
