+++ src/test/java/com/ignfab/minalac/generator/parameters/heightmaps/ReadableHeightmapParamsDeserializerTest.java
@@ -5,9 +5,7 @@ import org.junit.jupiter.api.Test;
 
 import com.ignfab.minalac.generator.parameters.ParamsTester;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertInstanceOf;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class ReadableHeightmapParamsDeserializerTest {
