+++ src/test/java/com/ignfab/minalac/generator/parameters/models/filters/ModelFilterMetadataGreaterThanParamsTest.java
@@ -4,9 +4,7 @@ import org.junit.jupiter.api.Test;
 
 import com.ignfab.minalac.generator.models.filters.ModelFilterOnMetadataValue;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertInstanceOf;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class ModelFilterMetadataGreaterThanParamsTest {
