+++ src/test/java/com/ignfab/minalac/generator/parameters/placeables/patterns/RepeatPatternParamsTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 
 import com.ignfab.minalac.generator.parameters.ParamsTester;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class RepeatPatternParamsTest {
     @Test
