+++ src/test/java/com/ignfab/minalac/generator/parameters/providers/OverpassProviderParamsTest.java
@@ -7,8 +7,8 @@ import org.geotools.referencing.CRS;
 import org.junit.jupiter.api.Test;
 
 import com.ignfab.minalac.generator.generation.Generation;
-import com.ignfab.minalac.generator.outputs.testing.TestingVoxelWorld;
 import com.ignfab.minalac.generator.utils.random.TestingSeed;
+import com.ignfab.minalac.generator.world.TestingVoxelWorld;
 
 import static org.junit.jupiter.api.Assertions.*;
