+++ src/test/java/com/ignfab/minalac/generator/parameters/tasks/ScheduleParamsTester.java
@@ -4,8 +4,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.fail;
+import static org.junit.jupiter.api.Assertions.*;
 
 public final class ScheduleParamsTester {
     // Utility class
