+++ src/test/java/com/ignfab/minalac/generator/world/TestingVoxelTile.java
@@ -1,11 +1,9 @@
-package com.ignfab.minalac.generator.outputs.testing;
+package com.ignfab.minalac.generator.world;
 
 
 import com.ignfab.minalac.generator.placeables.Placeable;
 import com.ignfab.minalac.generator.utils.world3d.WorldBBox3d;
 import com.ignfab.minalac.generator.utils.world3d.WorldCoords3d;
-import com.ignfab.minalac.generator.world.MapWriteException;
-import com.ignfab.minalac.generator.world.VoxelTile;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -13,43 +11,30 @@ import static org.junit.jupiter.api.Assertions.*;
  * Testing purpose {@link VoxelTile} intended to be used in unit tests.
  */
 public class TestingVoxelTile extends VoxelTile {
-    private String[] voxels;
+    private final String[] voxels;
 
     /**
      * A default {@code TestingVoxelTile} instance to pass as argument where it won't actually be used.
      */
-    public static final TestingVoxelTile UNUSED = new TestingVoxelTile(TestingVoxelWorld.UNUSED, WorldBBox3d.EMPTY);
+    public static final TestingVoxelTile UNUSED = new TestingVoxelTile(WorldBBox3d.EMPTY);
 
     /**
      * Constructs a new TestingVoxelTile.
-     *
-     * @param world World of which this tile is a part
-     * @param limits Limits of the voxel world to create
-     *
+     * <p>
      * Beware: Avoid large limits!
      * Each voxel is stored in memory as a string, which could be very large.
+     *
+     * @param limits Limits of the voxel world to create
      */
-    public TestingVoxelTile(TestingVoxelWorld world, WorldBBox3d limits) {
+    public TestingVoxelTile(WorldBBox3d limits) {
         super(limits);
         voxels = new String[limits.size().volume()];
     }
 
-    /**
-     * Constructs a new TestingVoxelTile and its corresponding world.
-     *
-     * @param limits Limits of the voxel world to create
-     *
-     * Beware: Avoid large limits!
-     * Each voxel is stored in memory as a string, which could be very large.
-     */
-    public TestingVoxelTile(WorldBBox3d limits) {
-        this(new TestingVoxelWorld(), limits);
-    }
-
     @Override
-    public void save() throws MapWriteException {}
+    public void save() {}
 
-    // This method should not be called with out of bounds coordinate
+    // This method should not be called with out-of-bounds coordinate
     private int index(int x, int y, int z) {
         return x - limits().minX() + limits().sizeX()
             * (y - limits().minY() + limits().sizeY()
