+++ src/test/java/com/ignfab/minalac/generator/world/VoxelWorldTest.java
@@ -6,8 +6,6 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import com.ignfab.minalac.generator.outputs.testing.TestingVoxel;
-import com.ignfab.minalac.generator.outputs.testing.TestingVoxelTile;
 import com.ignfab.minalac.generator.placeables.Placeable;
 import com.ignfab.minalac.generator.utils.world2d.WorldBBox2d;
 import com.ignfab.minalac.generator.utils.world3d.WorldBBox3d;
@@ -19,29 +17,23 @@ public class VoxelWorldTest {
     @Test
     public void testSetLimits() {
         VoxelWorld world = new VoxelWorldMock();
-        assertThrows(IllegalArgumentException.class, () -> {
-            world.setLimits(new WorldBBox3d(
-                    new WorldCoords3d(-21, -19, -18),
-                    new WorldCoords3d(21, 22, 23)
-            ));
-        });
+        assertThrows(IllegalArgumentException.class, () -> world.setLimits(new WorldBBox3d(
+            new WorldCoords3d(-21, -19, -18),
+            new WorldCoords3d(21, 22, 23)
+        )));
 
-        assertDoesNotThrow(() -> {
-            world.setLimits(new WorldBBox3d(
-                new WorldCoords3d(-1, -2, -3),
-                new WorldCoords3d(4, 5, 6))
-            );
-        });
+        assertDoesNotThrow(() -> world.setLimits(new WorldBBox3d(
+            new WorldCoords3d(-1, -2, -3),
+            new WorldCoords3d(4, 5, 6))
+        ));
 
         assertEquals(new WorldBBox3d(new WorldCoords3d(-1, -2, -3), new WorldCoords3d(4, 5, 6)), world.limits());
         assertEquals(new WorldCoords3d(2, 2, 2), world.getMetadata().getSpawn());
 
-        assertThrows(IllegalStateException.class, () -> {
-            world.setLimits(new WorldBBox3d(
-                new WorldCoords3d(-1, -2, -3),
-                new WorldCoords3d(7, 8, 9))
-            );
-        });
+        assertThrows(IllegalStateException.class, () -> world.setLimits(new WorldBBox3d(
+            new WorldCoords3d(-1, -2, -3),
+            new WorldCoords3d(7, 8, 9))
+        ));
     }
 
     @Test
@@ -60,35 +52,29 @@ public class VoxelWorldTest {
         b.place(tile, 1, -2, 1);
         c.place(tile, 1, -2, -2);
 
-        assertDoesNotThrow(() -> {
-            assertIterableEquals(
-                List.of(
-                    new PlacedVoxel(a, new WorldCoords3d(-1, 2, 6)),
-                    new PlacedVoxel(b, new WorldCoords3d(-1, 2, 2)),
-                    new PlacedVoxel(c, new WorldCoords3d(-1, 2, 0)),
-                    new PlacedVoxel(a, new WorldCoords3d(-1, 2, -4)),
-                    new PlacedVoxel(c, new WorldCoords3d(-1, 2, -5))
-                ),
-                tile.voxels(-1, 2)
-            );
-        });
+        assertDoesNotThrow(() -> assertIterableEquals(
+            List.of(
+                new PlacedVoxel(a, new WorldCoords3d(-1, 2, 6)),
+                new PlacedVoxel(b, new WorldCoords3d(-1, 2, 2)),
+                new PlacedVoxel(c, new WorldCoords3d(-1, 2, 0)),
+                new PlacedVoxel(a, new WorldCoords3d(-1, 2, -4)),
+                new PlacedVoxel(c, new WorldCoords3d(-1, 2, -5))
+            ),
+            tile.voxels(-1, 2)
+        ));
 
-        assertDoesNotThrow(() -> {
-            assertIterableEquals(
-                List.of(
-                    new PlacedVoxel(b, new WorldCoords3d(1, -2, 1)),
-                    new PlacedVoxel(c, new WorldCoords3d(1, -2, -2))
-                ),
-                tile.voxels(1, -2)
-            );
-        });
+        assertDoesNotThrow(() -> assertIterableEquals(
+            List.of(
+                new PlacedVoxel(b, new WorldCoords3d(1, -2, 1)),
+                new PlacedVoxel(c, new WorldCoords3d(1, -2, -2))
+            ),
+            tile.voxels(1, -2)
+        ));
 
-        assertDoesNotThrow(() -> {
-            assertIterableEquals(
-                Collections.emptyList(),
-                tile.voxels(0, -2)
-            );
-        });
+        assertDoesNotThrow(() -> assertIterableEquals(
+            Collections.emptyList(),
+            tile.voxels(0, -2)
+        ));
     }
 
     private static class VoxelWorldMock extends VoxelWorld {
@@ -106,10 +92,10 @@ public class VoxelWorldTest {
         }
 
         @Override
-        public void initialize() throws MapWriteException {}
+        public void initialize() {}
 
         @Override
-        public void finalizeAndSave() throws MapWriteException {}
+        public void finalizeAndSave() {}
 
         @Override
         public VoxelTile newTile(WorldBBox3d limits) {
